Hide `runToCompletion` behind `run(Write|Read)Operation`
This is the first step toward simplifying the underlying implementation.
