+++ subprojects/instant-execution/src/test/kotlin/org/gradle/instantexecution/serialization/codecs/AbstractUserTypeCodecTest.kt
@@ -18,7 +18,6 @@ package org.gradle.instantexecution.serialization.codecs
 
 import com.nhaarman.mockitokotlin2.mock
 import org.gradle.cache.internal.TestCrossBuildInMemoryCacheFactory
-import org.gradle.instantexecution.coroutines.runToCompletion
 import org.gradle.instantexecution.extensions.uncheckedCast
 import org.gradle.instantexecution.problems.ProblemsListener
 import org.gradle.instantexecution.problems.PropertyProblem
@@ -28,6 +27,8 @@ import org.gradle.instantexecution.serialization.DefaultWriteContext
 import org.gradle.instantexecution.serialization.IsolateOwner
 import org.gradle.instantexecution.serialization.MutableIsolateContext
 import org.gradle.instantexecution.serialization.beans.BeanConstructors
+import org.gradle.instantexecution.serialization.runReadOperation
+import org.gradle.instantexecution.serialization.runWriteOperation
 import org.gradle.instantexecution.serialization.withIsolate
 import org.gradle.internal.io.NullOutputStream
 import org.gradle.internal.serialize.Encoder
@@ -91,7 +92,7 @@ abstract class AbstractUserTypeCodecTest {
     ) {
         writeContextFor(KryoBackedEncoder(outputStream), codec, problemsListener).useToRun {
             withIsolateMock(codec) {
-                runToCompletion {
+                runWriteOperation {
                     write(graph)
                 }
             }
@@ -107,7 +108,7 @@ abstract class AbstractUserTypeCodecTest {
         readContextFor(inputStream, codec).run {
             initClassLoader(javaClass.classLoader)
             withIsolateMock(codec) {
-                runToCompletion {
+                runReadOperation {
                     read()
                 }
             }
@@ -140,7 +141,7 @@ abstract class AbstractUserTypeCodecTest {
             problemsListener = mock()
         )
 
-    protected
+    private
     fun userTypesCodec() = codecs().userTypesCodec
 
     protected
